+++ CHANGELOG.md
@@ -33,14 +33,17 @@ This release improves the WebradioDB integration and removes the radiobrowser.in
 
 ***
 
-## myMPD 16.0.2 (not yet released)
+## myMPD 16.1.0 (not yet released)
 
 This is a small bug fix release.
 
 ### Changelog
 
+- Feat: New API method MYMPD_API_SCRIPT_RELOAD to reload scripts from disk #1327
+- Upd: zh-Hans translation
 - Fix: Implicit conversion creating keys on systems with OpenSSL 1.1.x #1323
 - Fix: Request stored playlist entries in chunks (MPD 0.24)
+- Fix: Add missing event parameter to mympd-script #1325
 
 ***